requirements.txt
whisper
torch
coqui-tts
ollama
